refactor: moving slider_widget and inverting dependencies.
